# Reduce allocations in UnpackDomainName by better sizing slice