--- a/msg.go
+++ b/msg.go
@@ -42,7 +42,9 @@ const (
 	// maximum label length being 63. The wire format requires one extra byte over
 	// the presentation format, reducing the number of octets by 1. Each label in
 	// the name will be separated by a single period, with each octet in the label
-	// expanding to at most 4 bytes (\DDD).
+	// expanding to at most 4 bytes (\DDD). If all other labels are of the maximum
+	// length, then the final label can only be 61 octets long to not exceed the
+	// maximum allowed wire length.
 	maxDomainNamePresentationLength = 61*4 + 1 + 63*4 + 1 + 63*4 + 1 + 63*4 + 1
 )