This simple change greatly improves the performance and memory consumption of `UnpackDomainName`. It's perfectly safe to size `s` this large as it never escapes and will be stack allocated.

After this PR, `UnpackDomainName` will never be the source of memory garbage.

In terms of number of allocations and allocated memory, this is absolutely the best possible results, *i.e.* `UnpackDomainName` can never be more efficient with regards to memory usage.

```
name                                   old time/op    new time/op    delta
UnpackDomainName-12                       169ns ±13%     167ns ±19%     ~     (p=0.955 n=10+10)
UnpackDomainNameUnprintable-12            150ns ±13%     160ns ± 6%   +6.61%  (p=0.015 n=10+10)
UnpackDomainNameLongest-12               1.45µs ±30%    0.74µs ± 7%  -48.74%  (p=0.000 n=10+9)
UnpackDomainNameLongestUnprintable-12    6.05µs ±16%    3.21µs ± 7%  -47.00%  (p=0.000 n=10+10)

name                                   old alloc/op   new alloc/op   delta
UnpackDomainName-12                       64.0B ± 0%     64.0B ± 0%     ~     (all equal)
UnpackDomainNameUnprintable-12            48.0B ± 0%     48.0B ± 0%     ~     (all equal)
UnpackDomainNameLongest-12                 640B ± 0%      256B ± 0%  -60.00%  (p=0.000 n=10+10)
UnpackDomainNameLongestUnprintable-12    2.94kB ± 0%    1.02kB ± 0%  -65.22%  (p=0.000 n=10+10)

name                                   old allocs/op  new allocs/op  delta
UnpackDomainName-12                        1.00 ± 0%      1.00 ± 0%     ~     (all equal)
UnpackDomainNameUnprintable-12             1.00 ± 0%      1.00 ± 0%     ~     (all equal)
UnpackDomainNameLongest-12                 3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.000 n=10+10)
UnpackDomainNameLongestUnprintable-12      5.00 ± 0%      1.00 ± 0%  -80.00%  (p=0.000 n=10+10)
```

/cc @semihalev (for https://github.com/miekg/dns/issues/836#issuecomment-442413416 & https://github.com/miekg/dns/pull/839#discussion_r237073104)